+++ src/TimeoutMigrationTool.Tests/FakeToolState.cs
@@ -1,6 +1,5 @@
 ﻿namespace TimeoutMigrationTool.Tests
 {
-    using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
@@ -39,29 +38,6 @@
             var pendingBatch = Batches.First(x => x.State == BatchState.Pending);
 
             return Task.FromResult<BatchInfo>(pendingBatch);
-        }
-    }
-
-
-    public class FakeBatchInfo : BatchInfo
-    {
-        public FakeBatchInfo(int number, BatchState state, int numberOfTimeouts) : base(number, state, numberOfTimeouts)
-        {
-
-        }
-
-        public string[] TimeoutIds { get; set; }
-        public DateTime CutoffDate { get; set; }
-        public string EndpointName { get; set; }
-
-        public void MarkAsCompleted()
-        {
-            State = BatchState.Completed;
-        }
-
-        public void MarkAsStaged()
-        {
-            State = BatchState.Staged;
         }
     }
 }
